+++ src/v1/instance_admin_client.js
@@ -17,7 +17,6 @@
 const gapicConfig = require('./instance_admin_client_config.json');
 const gax = require('google-gax');
 const path = require('path');
-const protobuf = require('protobufjs');
 
 const VERSION = require('../../../package.json').version;
 
@@ -79,6 +78,16 @@ class InstanceAdminClient {
     opts = opts || {};
     this._descriptors = {};
 
+    if (global.isBrowser) {
+      // If we're in browser, we use gRPC fallback.
+      opts.fallback = true;
+    }
+
+    // If we are in browser, we are already using fallback because of the
+    // "browser" field in package.json.
+    // But if we were explicitly requested to use fallback, let's do it now.
+    const gaxModule = !global.isBrowser && opts.fallback ? gax.fallback : gax;
+
     const servicePath =
       opts.servicePath || opts.apiEndpoint || this.constructor.servicePath;
 
@@ -95,72 +104,80 @@ class InstanceAdminClient {
     // Create a `gaxGrpc` object, with any grpc-specific options
     // sent to the client.
     opts.scopes = this.constructor.scopes;
-    const gaxGrpc = new gax.GrpcClient(opts);
+    const gaxGrpc = new gaxModule.GrpcClient(opts);
 
     // Save the auth object to the client, for use by other methods.
     this.auth = gaxGrpc.auth;
 
     // Determine the client header string.
-    const clientHeader = [
-      `gl-node/${process.version}`,
-      `grpc/${gaxGrpc.grpcVersion}`,
-      `gax/${gax.version}`,
-      `gapic/${VERSION}`,
-    ];
+    const clientHeader = [];
+
+    if (typeof process !== 'undefined' && 'versions' in process) {
+      clientHeader.push(`gl-node/${process.versions.node}`);
+    }
+    clientHeader.push(`gax/${gaxModule.version}`);
+    if (opts.fallback) {
+      clientHeader.push(`gl-web/${gaxModule.version}`);
+    } else {
+      clientHeader.push(`grpc/${gaxGrpc.grpcVersion}`);
+    }
+    clientHeader.push(`gapic/${VERSION}`);
     if (opts.libName && opts.libVersion) {
       clientHeader.push(`${opts.libName}/${opts.libVersion}`);
     }
 
     // Load the applicable protos.
+    // For Node.js, pass the path to JSON proto file.
+    // For browsers, pass the JSON content.
+
+    const nodejsProtoPath = path.join(
+      __dirname,
+      '..',
+      '..',
+      'protos',
+      'protos.json'
+    );
     const protos = gaxGrpc.loadProto(
-      path.join(__dirname, '..', '..', 'protos'),
-      ['google/spanner/admin/instance/v1/spanner_instance_admin.proto']
+      opts.fallback ? require('../../protos/protos.json') : nodejsProtoPath
     );
 
     // This API contains "path templates"; forward-slash-separated
     // identifiers to uniquely identify resources within the API.
     // Create useful helper objects for these.
     this._pathTemplates = {
-      instancePathTemplate: new gax.PathTemplate(
+      instancePathTemplate: new gaxModule.PathTemplate(
         'projects/{project}/instances/{instance}'
       ),
-      instanceConfigPathTemplate: new gax.PathTemplate(
+      instanceConfigPathTemplate: new gaxModule.PathTemplate(
         'projects/{project}/instanceConfigs/{instance_config}'
       ),
-      projectPathTemplate: new gax.PathTemplate('projects/{project}'),
+      projectPathTemplate: new gaxModule.PathTemplate('projects/{project}'),
     };
 
     // Some of the methods on this service return "paged" results,
     // (e.g. 50 results at a time, with tokens to get subsequent
     // pages). Denote the keys used for pagination and results.
     this._descriptors.page = {
-      listInstanceConfigs: new gax.PageDescriptor(
+      listInstanceConfigs: new gaxModule.PageDescriptor(
         'pageToken',
         'nextPageToken',
         'instanceConfigs'
       ),
-      listInstances: new gax.PageDescriptor(
+      listInstances: new gaxModule.PageDescriptor(
         'pageToken',
         'nextPageToken',
         'instances'
       ),
     };
-    let protoFilesRoot = new gax.GoogleProtoFilesRoot();
-    protoFilesRoot = protobuf.loadSync(
-      path.join(
-        __dirname,
-        '..',
-        '..',
-        'protos',
-        'google/spanner/admin/instance/v1/spanner_instance_admin.proto'
-      ),
-      protoFilesRoot
-    );
+
+    const protoFilesRoot = opts.fallback
+      ? gaxModule.protobuf.Root.fromJSON(require('../../protos/protos.json'))
+      : gaxModule.protobuf.loadSync(nodejsProtoPath);
 
     // This API contains "long-running operations", which return a
     // an Operation object that allows for tracking of the operation,
     // rather than holding a request open.
-    this.operationsClient = new gax.lro({
+    this.operationsClient = new gaxModule.lro({
       auth: gaxGrpc.auth,
       grpc: gaxGrpc.grpc,
     }).operationsClient(opts);
@@ -179,12 +196,12 @@ class InstanceAdminClient {
     );
 
     this._descriptors.longrunning = {
-      createInstance: new gax.LongrunningDescriptor(
+      createInstance: new gaxModule.LongrunningDescriptor(
         this.operationsClient,
         createInstanceResponse.decode.bind(createInstanceResponse),
         createInstanceMetadata.decode.bind(createInstanceMetadata)
       ),
-      updateInstance: new gax.LongrunningDescriptor(
+      updateInstance: new gaxModule.LongrunningDescriptor(
         this.operationsClient,
         updateInstanceResponse.decode.bind(updateInstanceResponse),
         updateInstanceMetadata.decode.bind(updateInstanceMetadata)
@@ -207,7 +224,9 @@ class InstanceAdminClient {
     // Put together the "service stub" for
     // google.spanner.admin.instance.v1.InstanceAdmin.
     const instanceAdminStub = gaxGrpc.createStub(
-      protos.google.spanner.admin.instance.v1.InstanceAdmin,
+      opts.fallback
+        ? protos.lookupService('google.spanner.admin.instance.v1.InstanceAdmin')
+        : protos.google.spanner.admin.instance.v1.InstanceAdmin,
       opts
     );
 
@@ -226,18 +245,16 @@ class InstanceAdminClient {
       'testIamPermissions',
     ];
     for (const methodName of instanceAdminStubMethods) {
-      this._innerApiCalls[methodName] = gax.createApiCall(
-        instanceAdminStub.then(
-          stub =>
-            function() {
-              const args = Array.prototype.slice.call(arguments, 0);
-              return stub[methodName].apply(stub, args);
-            },
-          err =>
-            function() {
-              throw err;
-            }
-        ),
+      const innerCallPromise = instanceAdminStub.then(
+        stub => (...args) => {
+          return stub[methodName].apply(stub, args);
+        },
+        err => () => {
+          throw err;
+        }
+      );
+      this._innerApiCalls[methodName] = gaxModule.createApiCall(
+        innerCallPromise,
         defaults[methodName],
         this._descriptors.page[methodName] ||
           this._descriptors.longrunning[methodName]