+++ src/v1/spanner_client.js
@@ -61,6 +61,16 @@ class SpannerClient {
     opts = opts || {};
     this._descriptors = {};
 
+    if (global.isBrowser) {
+      // If we're in browser, we use gRPC fallback.
+      opts.fallback = true;
+    }
+
+    // If we are in browser, we are already using fallback because of the
+    // "browser" field in package.json.
+    // But if we were explicitly requested to use fallback, let's do it now.
+    const gaxModule = !global.isBrowser && opts.fallback ? gax.fallback : gax;
+
     const servicePath =
       opts.servicePath || opts.apiEndpoint || this.constructor.servicePath;
 
@@ -77,36 +87,51 @@ class SpannerClient {
     // Create a `gaxGrpc` object, with any grpc-specific options
     // sent to the client.
     opts.scopes = this.constructor.scopes;
-    const gaxGrpc = new gax.GrpcClient(opts);
+    const gaxGrpc = new gaxModule.GrpcClient(opts);
 
     // Save the auth object to the client, for use by other methods.
     this.auth = gaxGrpc.auth;
 
     // Determine the client header string.
-    const clientHeader = [
-      `gl-node/${process.version}`,
-      `grpc/${gaxGrpc.grpcVersion}`,
-      `gax/${gax.version}`,
-      `gapic/${VERSION}`,
-    ];
+    const clientHeader = [];
+
+    if (typeof process !== 'undefined' && 'versions' in process) {
+      clientHeader.push(`gl-node/${process.versions.node}`);
+    }
+    clientHeader.push(`gax/${gaxModule.version}`);
+    if (opts.fallback) {
+      clientHeader.push(`gl-web/${gaxModule.version}`);
+    } else {
+      clientHeader.push(`grpc/${gaxGrpc.grpcVersion}`);
+    }
+    clientHeader.push(`gapic/${VERSION}`);
     if (opts.libName && opts.libVersion) {
       clientHeader.push(`${opts.libName}/${opts.libVersion}`);
     }
 
     // Load the applicable protos.
+    // For Node.js, pass the path to JSON proto file.
+    // For browsers, pass the JSON content.
+
+    const nodejsProtoPath = path.join(
+      __dirname,
+      '..',
+      '..',
+      'protos',
+      'protos.json'
+    );
     const protos = gaxGrpc.loadProto(
-      path.join(__dirname, '..', '..', 'protos'),
-      ['google/spanner/v1/spanner.proto']
+      opts.fallback ? require('../../protos/protos.json') : nodejsProtoPath
     );
 
     // This API contains "path templates"; forward-slash-separated
     // identifiers to uniquely identify resources within the API.
     // Create useful helper objects for these.
     this._pathTemplates = {
-      databasePathTemplate: new gax.PathTemplate(
+      databasePathTemplate: new gaxModule.PathTemplate(
         'projects/{project}/instances/{instance}/databases/{database}'
       ),
-      sessionPathTemplate: new gax.PathTemplate(
+      sessionPathTemplate: new gaxModule.PathTemplate(
         'projects/{project}/instances/{instance}/databases/{database}/sessions/{session}'
       ),
     };
@@ -115,7 +140,7 @@ class SpannerClient {
     // (e.g. 50 results at a time, with tokens to get subsequent
     // pages). Denote the keys used for pagination and results.
     this._descriptors.page = {
-      listSessions: new gax.PageDescriptor(
+      listSessions: new gaxModule.PageDescriptor(
         'pageToken',
         'nextPageToken',
         'sessions'
@@ -125,10 +150,12 @@ class SpannerClient {
     // Some of the methods on this service provide streaming responses.
     // Provide descriptors for these.
     this._descriptors.stream = {
-      executeStreamingSql: new gax.StreamDescriptor(
+      executeStreamingSql: new gaxModule.StreamDescriptor(
         gax.StreamType.SERVER_STREAMING
       ),
-      streamingRead: new gax.StreamDescriptor(gax.StreamType.SERVER_STREAMING),
+      streamingRead: new gaxModule.StreamDescriptor(
+        gax.StreamType.SERVER_STREAMING
+      ),
     };
 
     // Put together the default options sent with requests.
@@ -147,7 +174,9 @@ class SpannerClient {
     // Put together the "service stub" for
     // google.spanner.v1.Spanner.
     const spannerStub = gaxGrpc.createStub(
-      protos.google.spanner.v1.Spanner,
+      opts.fallback
+        ? protos.lookupService('google.spanner.v1.Spanner')
+        : protos.google.spanner.v1.Spanner,
       opts
     );
 
@@ -155,6 +184,7 @@ class SpannerClient {
     // and create an API call method for each.
     const spannerStubMethods = [
       'createSession',
+      'batchCreateSessions',
       'getSession',
       'listSessions',
       'deleteSession',
@@ -170,18 +200,16 @@ class SpannerClient {
       'partitionRead',
     ];
     for (const methodName of spannerStubMethods) {
-      this._innerApiCalls[methodName] = gax.createApiCall(
-        spannerStub.then(
-          stub =>
-            function() {
-              const args = Array.prototype.slice.call(arguments, 0);
-              return stub[methodName].apply(stub, args);
-            },
-          err =>
-            function() {
-              throw err;
-            }
-        ),
+      const innerCallPromise = spannerStub.then(
+        stub => (...args) => {
+          return stub[methodName].apply(stub, args);
+        },
+        err => () => {
+          throw err;
+        }
+      );
+      this._innerApiCalls[methodName] = gaxModule.createApiCall(
+        innerCallPromise,
         defaults[methodName],
         this._descriptors.page[methodName] ||
           this._descriptors.stream[methodName]
@@ -309,6 +337,74 @@ class SpannerClient {
     });
 
     return this._innerApiCalls.createSession(request, options, callback);
+  }
+
+  /**
+   * Creates multiple new sessions.
+   *
+   * This API can be used to initialize a session cache on the clients.
+   * See https://goo.gl/TgSFN2 for best practices on session cache management.
+   *
+   * @param {Object} request
+   *   The request object that will be sent.
+   * @param {string} request.database
+   *   Required. The database in which the new sessions are created.
+   * @param {Object} [request.sessionTemplate]
+   *   Parameters to be applied to each created session.
+   *
+   *   This object should have the same structure as [Session]{@link google.spanner.v1.Session}
+   * @param {number} [request.sessionCount]
+   *   Required. The number of sessions to be created in this batch call.
+   *   The API may return fewer than the requested number of sessions. If a
+   *   specific number of sessions are desired, the client can make additional
+   *   calls to BatchCreateSessions (adjusting
+   *   session_count
+   *   as necessary).
+   * @param {Object} [options]
+   *   Optional parameters. You can override the default settings for this call, e.g, timeout,
+   *   retries, paginations, etc. See [gax.CallOptions]{@link https://googleapis.github.io/gax-nodejs/interfaces/CallOptions.html} for the details.
+   * @param {function(?Error, ?Object)} [callback]
+   *   The function which will be called with the result of the API call.
+   *
+   *   The second parameter to the callback is an object representing [BatchCreateSessionsResponse]{@link google.spanner.v1.BatchCreateSessionsResponse}.
+   * @returns {Promise} - The promise which resolves to an array.
+   *   The first element of the array is an object representing [BatchCreateSessionsResponse]{@link google.spanner.v1.BatchCreateSessionsResponse}.
+   *   The promise has a method named "cancel" which cancels the ongoing API call.
+   *
+   * @example
+   *
+   * const spanner = require('@google-cloud/spanner');
+   *
+   * const client = new spanner.v1.SpannerClient({
+   *   // optional auth parameters.
+   * });
+   *
+   * const formattedDatabase = client.databasePath('[PROJECT]', '[INSTANCE]', '[DATABASE]');
+   * client.batchCreateSessions({database: formattedDatabase})
+   *   .then(responses => {
+   *     const response = responses[0];
+   *     // doThingsWith(response)
+   *   })
+   *   .catch(err => {
+   *     console.error(err);
+   *   });
+   */
+  batchCreateSessions(request, options, callback) {
+    if (options instanceof Function && callback === undefined) {
+      callback = options;
+      options = {};
+    }
+    request = request || {};
+    options = options || {};
+    options.otherArgs = options.otherArgs || {};
+    options.otherArgs.headers = options.otherArgs.headers || {};
+    options.otherArgs.headers[
+      'x-goog-request-params'
+    ] = gax.routingHeader.fromParams({
+      database: request.database,
+    });
+
+    return this._innerApiCalls.batchCreateSessions(request, options, callback);
   }
 
   /**
@@ -616,9 +712,6 @@ class SpannerClient {
    * @param {string} request.sql
    *   Required. The SQL string.
    * @param {Object} [request.transaction]
-   *   The transaction to use. If none is provided, the default is a
-   *   temporary read-only transaction with strong concurrency.
-   *
    *   The transaction to use.
    *
    *   For queries, if none is provided, the default is a temporary read-only
@@ -756,9 +849,6 @@ class SpannerClient {
    * @param {string} request.sql
    *   Required. The SQL string.
    * @param {Object} [request.transaction]
-   *   The transaction to use. If none is provided, the default is a
-   *   temporary read-only transaction with strong concurrency.
-   *
    *   The transaction to use.
    *
    *   For queries, if none is provided, the default is a temporary read-only
@@ -875,8 +965,9 @@ class SpannerClient {
    *
    * Statements are executed in order, sequentially.
    * ExecuteBatchDmlResponse will contain a
-   * ResultSet for each DML statement that has successfully executed. If a
-   * statement fails, its error status will be returned as part of the
+   * ResultSet for each DML statement that has
+   * successfully executed. If a statement fails, its error status will be
+   * returned as part of the
    * ExecuteBatchDmlResponse. Execution will
    * stop at the first failed statement; the remaining statements will not run.
    *